+++ packages/react/src/PrivacyPolicy/index.tsx
@@ -1,9 +1,9 @@
 import { type PrivacyPolicyProps } from "@incubateur-ademe/legal-pages-markdown";
 import { type ReactNode, useId } from "react";
 
-import { ClientOnly } from "./ClientOnly";
-import { ClientPortal } from "./ClientPortal";
-import { htmlParser } from "./htmlParser";
+import { ClientOnly } from "../utils/ClientOnly";
+import { ClientPortal } from "../utils/ClientPortal";
+import { htmlParser } from "../utils/htmlParser";
 
 export { type PrivacyPolicyProps };
 
@@ -109,3 +109,5 @@ export const PrivacyPolicy = async ({
     </>
   );
 };
+
+export default PrivacyPolicy;