+++ packages/react/src/types.ts
@@ -1,5 +1,0 @@
-export interface CommonProps {
-  date?: string;
-  includeBetaGouv?: boolean;
-  siteName: string;
-}